transform helper (#56)
* adding helper function to utils.py

The helper function parses the run properties and results data so that the user doesn't need to include any parsing script in their transform script. The helper function only accepts a user-defined function that accepts an array of results data as its input and returns an array of results data matching the assay design it is used with. Also, the helper function requires that users explicitly define the run properties in their transform script, otherwise the labkey transform script parser does not recognize ${runInfo}.

* Create example transform script that uses labkey.utils.transform_helper

* Delete assay_transform_function_example.py

moving filter to /docs folder

* documentation for transform_helper function

* Update transform_helper.md

* Update transform_helper.md

* updating transform_helper to be more robust and use snake case

switched from camel case to snake case for variable naming. additionally, changed script to get file_path_in from the row beginning with "runDataUploadedFile" in runProperties file. This is to better handle results data with numerical header names, such as when the results data is a grid representing well plates.

* Update labkey/utils.py

Co-authored-by: Alan Vezina <alanv@labkey.com>

* Update utils.py

using with statements when opening files

---------

Co-authored-by: Alan Vezina <alanv@labkey.com>